requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Machine Learning
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
joblib==1.3.2

# MLflow
mlflow<2.10
psycopg2-binary==2.9.9

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0

# Authentication & Database
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
bcrypt==4.1.2

# Data Version Control
dvc==3.55.2
dvc-s3==3.2.0

# Development & Linting
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

matplotlib==3.8.2
seaborn==0.13.1
